+++ packages/client/src/network/http/friend.ts
@@ -1,0 +1,13 @@
+import { get } from './request/methods';
+
+/**
+ *
+ * @param userId
+ * @returns
+ */
+export function getFriends(userId?: string) {
+  return get({
+    url: '/friend/get',
+    data: { userId },
+  });
+}