+++ packages/client/src/network/http/request/methods.ts
@@ -3,8 +3,8 @@ import { request, RequestConfig, Response } from './request';
 /**
  * get 请求方法
  */
-export function get<REQCONF, RESDATA>(config: RequestConfig<REQCONF>) {
-  return request.request<Response<RESDATA>>({
+export function get<REQ_DATA, RES_DATA>(config: RequestConfig<REQ_DATA>) {
+  return request.request<Response<RES_DATA>>({
     ...config,
     method: 'GET',
     params: config.data,
@@ -14,8 +14,8 @@ export function get<REQCONF, RESDATA>(config: RequestConfig<REQCONF>) {
 /**
  * post 请求方法
  */
-export function post<REQCONF, RESDATA>(config: RequestConfig<REQCONF>) {
-  return request.request<Response<RESDATA>>({
+export function post<REQ_DATA, RES_DATA>(config: RequestConfig<REQ_DATA>) {
+  return request.request<Response<RES_DATA>>({
     ...config,
     method: 'POST',
   });