+++ packages/client/src/pages/home/index.tsx
@@ -1,7 +1,9 @@
 import { FC } from 'react';
+import { Route, Routes } from 'react-router-dom';
 import HomeLeftPanel from './leftPanel';
 import HomeMain from './main';
 import './index.less';
+import Preview from './main/components/preview';
 
 interface HomeProps {}
 
@@ -9,7 +11,10 @@ const Home: FC<HomeProps> = (props) => {
   return (
     <div className='home-page'>
       <HomeLeftPanel />
-      <HomeMain />
+      <Routes>
+        <Route path='/message/:id' element={<HomeMain />} />
+        <Route path='*' element={<Preview />} />
+      </Routes>
     </div>
   );
 };