+++ packages/client/src/store/index.ts
@@ -1,5 +1,3 @@
-import { observable, action } from 'mobx';
-
 class Store {}
 
 export default Store;