+++ packages/client/vite.config.ts
@@ -1,5 +1,7 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
+import { createSvgIconsPlugin } from 'vite-plugin-svg-icons';
+import * as path from 'path';
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -11,6 +13,12 @@ export default defineConfig({
           ['@babel/plugin-proposal-class-properties', { loose: true }],
         ],
       },
+    }),
+    createSvgIconsPlugin({
+      // 指定需要缓存的图标文件夹
+      iconDirs: [path.resolve(process.cwd(), 'src/icons')],
+      // 指定symbolId格式
+      symbolId: 'icon-[dir]-[name]',
     }),
   ],
 });