+++ packages/client/src/App.tsx
@@ -1,20 +1,18 @@
 import { FC, useEffect } from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { ConfigProvider } from 'antd';
-import Store from './store';
 import Login from './pages/login';
 import Home from './pages/home';
 import { WSManager } from './network/websocket';
 import { getBrandColor, changeThemeMode } from './utils/theme';
 import Theme from './pages/components/theme';
+import { useSystemThemeModeChange } from './hooks/useSystemThemeModeChange';
 // import 'antd/dist/antd.variable.css';
 // import 'antd/dist/antd.dark.css';
 import './static/style/.variable.less';
 import './static/style/.dark.less';
 import './static/style/variable.less';
 import './static/style/replenish.less';
-
-const store = new Store();
 
 const App: FC<{}> = () => {
   useEffect(() => {
@@ -23,6 +21,8 @@ const App: FC<{}> = () => {
 
     changeThemeMode();
   }, []);
+
+  useSystemThemeModeChange();
 
   ConfigProvider.config({
     theme: {