+++ packages/client/src/network/http/friend.ts
@@ -1,12 +1,12 @@
 import { get } from './request/methods';
 
 /**
- *
+ * 获取好友
  * @param userId
  * @returns
  */
 export function getFriends(userId?: string) {
-  return get({
+  return get<any, Friend[]>({
     url: '/friend/get',
     data: { userId },
   });