+++ packages/client/src/network/http/request/methods.ts
@@ -3,7 +3,7 @@ import { request, RequestConfig, Response } from './request';
 /**
  * get 请求方法
  */
-export function get<REQ_DATA, RES_DATA>(config: RequestConfig<REQ_DATA>) {
+export function get<REQ_DATA = any, RES_DATA = any>(config: RequestConfig<REQ_DATA>) {
   return request.request<Response<RES_DATA>>({
     ...config,
     method: 'GET',
@@ -14,7 +14,7 @@ export function get<REQ_DATA, RES_DATA>(config: RequestConfig<REQ_DATA>) {
 /**
  * post 请求方法
  */
-export function post<REQ_DATA, RES_DATA>(config: RequestConfig<REQ_DATA>) {
+export function post<REQ_DATA = any, RES_DATA = any>(config: RequestConfig<REQ_DATA>) {
   return request.request<Response<RES_DATA>>({
     ...config,
     method: 'POST',