+++ packages/client/src/pages/home/main/components/chat/index.tsx
@@ -1,0 +1,9 @@
+import { FC } from 'react';
+
+interface ChatProps {}
+
+const Chat: FC<ChatProps> = (props) => {
+  return <div>chat</div>;
+};
+
+export default Chat;