+++ packages/client/src/types/index.d.ts
@@ -4,3 +4,15 @@ declare interface UserInfo {
   name: string;
   avatar: string;
 }
+
+declare interface Friend {
+  id: number;
+
+  note?: string;
+
+  isTop?: boolean;
+
+  isBottom?: boolean;
+
+  info: UserInfo;
+}